+++ src/modules/localStorage.js
@@ -11,13 +11,15 @@ export default class Storage {
     }
     return tasks;
   }
-// -- When pressed add task we call this functin to add it to LS --
+
+  // -- When pressed add task we call this functin to add it to LS --
   static addTask(task) {
     const tasks = Storage.getTasks();
     tasks.push(task);
     localStorage.setItem('tasks', JSON.stringify(tasks));
   }
-// -- When pressed remove task we call this functin to remove it from LS --
+
+  // -- When pressed remove task we call this functin to remove it from LS --
   static removeTask(id) {
     let tasks = JSON.parse(localStorage.getItem('tasks'));
     tasks = tasks.filter((e) => e.index.toString() !== id.toString());
@@ -27,7 +29,8 @@ export default class Storage {
     }
     localStorage.setItem('tasks', JSON.stringify(tasks));
   }
-// -- When pressed checkbox, we change the status complete of the task  to true in LS --
+
+  // -- When pressed checkbox, we change the status complete of the task  to true in LS --
   static updateStatus(id) {
     const tasks = JSON.parse(localStorage.getItem('tasks'));
     for (let i = 0; i < tasks.length; i += 1) {
@@ -42,6 +45,7 @@ export default class Storage {
       }
     }
   }
+
   // -- Update Local Storage when mdoify task description --
   static editTask(newTitle, newDescription, newDate, id) {
     const tasks = JSON.parse(localStorage.getItem('tasks'));
